Adding capability to listen for mirrored notifications via pushbullet
Initial changes:
- Notification format changed to be more practical.
- Notifications dismiss themselves when they expire.
- Can now mirror notifications from android phones via pushbullet and
receive them in pogo-finder.
- Currently allows for location changes and step limit changes.
- Will re-login every 10 minutes so that the API doesn't stop sending
data.
- Notification respects the "ampm_clock" argument.
- When the API returns a negative time to disappear, we assume a 15
minute time limit. This is actually probably in the 10-15 minute range,
so 15 minutes max should be an appropriate assumption (except when the
app does its initial scan). This is indicated in the notification and on
the map with an "(!)" indication and a message.